+++ front/src/app/meta/run-visualizer.tsx
@@ -8,13 +8,12 @@ import {
   ModalHeader,
 } from "@heroui/react";
 import { SampleVisualizer } from "./sample-visualizer";
-import { RunCharts } from "./run-charts";
+import { RunMetrics } from "./run-metrics";
 
 export function RunVisualizer() {
   const focused = useFocusedRun();
 
   if (!focused?.samples.length) return null;
-  console.log(focused);
 
   return (
     <Modal
@@ -31,13 +30,11 @@ export function RunVisualizer() {
           </span>
         </ModalHeader>
         <ModalBody className="overflow-auto">
-          <div className="scrollbar-hide w-full snap-y space-y-1 gap-x-1 overflow-y-auto">
-            <h1 className="my-2 w-full text-center text-3xl">Charts</h1>
+          <div id="run-visualizer" className="scrollbar-hide w-full snap-y space-y-1 gap-x-1 overflow-y-auto">
+
+            <RunMetrics />
             <Divider className="mx-auto my-3 w-11/12 opacity-40" />
-            <RunCharts />
-            <Divider className="mx-auto my-3 w-11/12 opacity-40" />
-            <h1 className="my-2 w-full text-center text-3xl">Samples</h1>
-            <Divider className="mx-auto my-3 w-11/12 opacity-40" />
+
             {focused?.samples.map((sample) => (
               <SampleVisualizer
                 key={`${sample.runId}-${sample.id}`}
