+++ front/src/components/metrics/heatmap-chart.tsx
@@ -1,74 +1,13 @@
 import { Accordion } from "@/components/custom-accordion";
 import { useFocusedRun } from "@/hooks/use-focused-run";
-import { api } from "@/lib/axios";
 import {
-  infernoColorGenerator,
-  viridisColorGenerator,
+  infernoColorGenerator
 } from "@/utils/color-generator";
-import { queryKeys } from "@/utils/query-keys-factory";
-import { faker } from "@faker-js/faker";
-import { Switch, Tooltip as TooltipHero } from "@heroui/react";
-import { useQuery } from "@tanstack/react-query";
-import {
-  BarElement,
-  CategoryScale,
-  Chart as ChartJS,
-  Colors,
-  Legend,
-  LinearScale,
-  LogarithmicScale,
-  Title,
-  Tooltip,
-} from "chart.js";
+import { Tooltip as TooltipHero } from "@heroui/react";
 import interpolate from "color-interpolate";
 import { ChartNoAxesColumnIncreasing } from "lucide-react";
 import { useState } from "react";
-import { Bar } from "react-chartjs-2";
-
-ChartJS.register(
-  CategoryScale,
-  LinearScale,
-  LogarithmicScale,
-  BarElement,
-  Title,
-  Tooltip,
-  Legend,
-  Colors,
-);
-
-export function RunCharts() {
-  const focused = useFocusedRun();
-  const { data } = useQuery({
-    queryKey: queryKeys.getCharts(focused),
-    queryFn: async () => {
-      const response = await api.get(`/v1/metagenomics/${focused?.id}/charts`);
-      return response.data;
-    },
-  });
-
-  if (!data) return null;
-
-  return (
-    <>
-      <BarChart
-        title="Total reads per sample (Classified vs Unclassified)"
-        dataSets={data.viralDatasets}
-      />
-      <BarChart
-        title="Reads per family per sample (absolute)"
-        legend="Family"
-        dataSets={data.familyDatasets}
-      />
-      <HeatMapChart title="HeatMap" dataSets={data.familyDatasets} />
-    </>
-  );
-}
-
-interface ChartProps {
-  title: string;
-  legend?: string;
-  dataSets: { dataSetTitle: string; data: number[] }[];
-}
+import { ChartProps } from "./types";
 
 const colors = [
   infernoColorGenerator(1),
@@ -85,7 +24,9 @@ export function HeatMapChart({ title, dataSets }: ChartProps) {
 
   const processedDataSets = dataSets.map((dts) => ({
     ...dts,
-    data: dts.data.map((v) => (log && v > 0 ? Number(Math.log10(v).toFixed(2)) : v)),
+    data: dts.data.map((v) =>
+      log && v > 0 ? Number(Math.log10(v).toFixed(2)) : v,
+    ),
   }));
 
   const flatData = processedDataSets.flatMap((dts) => dts.data);
@@ -175,84 +116,6 @@ export function HeatMapChart({ title, dataSets }: ChartProps) {
               <span className="w-32 p-1" />
             </div>
           </div>
-        </div>
-      </div>
-    </Accordion>
-  );
-}
-
-export function BarChart({ title, dataSets, legend }: ChartProps) {
-  const [show, setShow] = useState(true);
-  const [log, setLog] = useState(false);
-
-  const focused = useFocusedRun();
-
-  const processedDataSets = dataSets.map((dts) => ({
-    ...dts,
-    data: dts.data.map((v) => (log ? Number(Math.log10(v).toFixed(2)) : v)),
-  }));
-
-  return (
-    <Accordion
-      show={show}
-      toggle={() => setShow(!show)}
-      title={title}
-      actions={[
-        {
-          label: "Log10",
-          active: log,
-          icon: <ChartNoAxesColumnIncreasing />,
-          onPress: () => setLog(!log),
-        },
-      ]}
-    >
-      <div className="flex h-full items-center justify-center">
-        <div className="flex h-[85vh] w-full items-center justify-center p-4">
-          <Bar
-            data={{
-              labels: focused?.samples.map((s) => s.name),
-              datasets: processedDataSets.map((d, i) => ({
-                label: d.dataSetTitle,
-                data: d.data,
-                backgroundColor: viridisColorGenerator(
-                  i / (processedDataSets.length - 1),
-                ),
-              })),
-            }}
-            options={{
-              responsive: true,
-              animation: { duration: 500 },
-              aspectRatio: 19.5 / 9,
-              scales: {
-                x: {
-                  stacked: true,
-                  title: {
-                    display: true,
-                    text: "Samples",
-                  },
-                },
-                y: {
-                  stacked: true,
-                  type: "linear",
-                  title: {
-                    display: true,
-                    text: "Reads",
-                  },
-                },
-              },
-
-              plugins: {
-                legend: {
-                  position: "right" as const,
-                  title: {
-                    display: true,
-                    text: legend,
-                    position: "center",
-                  },
-                },
-              },
-            }}
-          />
         </div>
       </div>
     </Accordion>
