+++ front/src/components/metrics/types.ts
@@ -1,0 +1,5 @@
+export interface ChartProps {
+  title: string;
+  legend?: string;
+  dataSets: { dataSetTitle: string; data: number[] }[];
+}
